Hosts draft.
